third_party/blink/web_tests/virtual/css-anchor-update-enabled/external/wpt/css/css-anchor-position/position-area-scrolling-003.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Initial scroll position
  assert_equals: e7 height expected 100 but got 350
[FAIL] Scroll to 40,60
  assert_equals: e7 height expected 100 but got 350
[FAIL] Reattach at 40,60
  assert_equals: e7 height expected 100 but got 350
[FAIL] Redisplay at 40,60
  assert_equals: e1 x expected 120 but got 100
[FAIL] Scroll to 0,0 and relayout
  assert_equals: e1 x expected 120 but got 100
[FAIL] Redisplay at 0,0
  assert_equals: e7 height expected 100 but got 350
Harness: the test ran to completion.
